feat(site): fix hydration error in tab mode hint
